+++ apps/demo/src/app/chuck-norris/chuck-norris.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ChuckNorrisComponent } from './chuck-norris.component';
+
+
+@NgModule({
+  declarations: [ChuckNorrisComponent],
+  imports: [
+    CommonModule
+  ]
+})
+export class ChuckNorrisModule { }