+++ apps/demo/src/app/xkcd/xkcd.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { XkcdComponent } from './xkcd.component';
+
+
+@NgModule({
+  declarations: [XkcdComponent],
+  imports: [
+    CommonModule
+  ]
+})
+export class XkcdModule { }