+++ libs/markdown/extended/public_api.ts
@@ -1,0 +1,2 @@
+export * from './markdown-extended.module';
+export * from './render-markdown-extended/render-markdown-extended.component';